pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bertopic-app"
version = "2.0.0"
description = "Modern BERTopic Streamlit Application with Model2Vec"
authors = [
    {name = "BERTopic Team", email = "team@bertopic.ai"}
]
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "streamlit>=1.39.0",
    "pandas>=2.2.3",
    "numpy>=2.1.3",
    "model2vec>=0.2.8",
    "bertopic>=0.17.0",
    "scikit-learn>=1.7.2",
    "umap-learn>=0.5.8",
    "hdbscan>=0.8.41",
    "plotly>=5.24.1",
    "safetensors>=0.6.2",
    "loguru>=0.7.3",
    "rich>=13.9.4",
    "aiofiles>=24.1.0",
    "tqdm>=4.67.1",
    "pydantic>=2.10.3",
    "psutil>=6.1.0",
    "pyyaml>=6.0.2",
    "httpx>=0.28.1",
    "orjson>=3.10.12",
    "python-dateutil>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "streamlit-profiler>=0.2.0",
    "memory-profiler>=0.61.0",
]

[project.urls]
homepage = "https://github.com/bertopic/bertopic-app"
repository = "https://github.com/bertopic/bertopic-app"
documentation = "https://bertopic-app.readthedocs.io"

[project.scripts]
bertopic-app = "bertopic_app.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=src/bertopic_app --cov-report=html --cov-report=term-missing"